add (disabled) code to mangle email/srpPW to match picl-gherkin bug
https://github.com/mozilla/picl-gherkin/issues/33 covers a bug in the JS
client library which incorrectly performs double-encoding of the account
email address and the derived srpPW fields. This results in
accounts (srpVerifier and unwrapBkey) that don't match the spec,
breaking interoperability with other (compliant) implementations.

By turning on "MANGLE=True" on line 198, this code will match the buggy
behavior of picl-gherkin, enabling temporary interoperability.

We are changing picl-gherkin to match the spec, of course, not the other
way around.